LPD-28762 Check dependency version numbers (node SF)
Match all projects' dependencies between them and show those that are not
the same. Then do the same with devDependencies.

I'm also refactoring checkPackageJSONFiles because it got too large and
difficult to read.
