--- modules/_node-scripts/preflight/checkPackageJSONFiles.mjs
@@ -1,627 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import fg from 'fast-glob';
-import fs from 'fs';
-import fsPromises from 'fs/promises';
-import path from 'path';
-import url from 'url';
-
-import {getRootDir} from '../util/constants.mjs';
-import projectScopeRequire from '../util/projectScopeRequire.mjs';
-
-/**
- * Runs checks against package.json files; detects:
- *
- * - Bad package names (ie. packages without named scopes).
- * - Banned dependencies.
- * - Incorrect npm scripts
- *
- * Returns a (possibly empty) array of error messages.
- */
-export async function checkPackageJSONFiles() {
-	let packages = await fg('**/package.json', {
-		ignore: [
-			'_node-scripts',
-			'**/build',
-			'**/classes',
-			'**/frontend-js-jquery-web',
-			'**/node_modules',
-			'**/osb-faro-theme',
-			'**/osb-faro-web',
-			'**/sdk',
-			'test',
-		],
-	});
-
-	// Filters out packages that have their own yarn.lock
-
-	packages = packages.filter((packagePath) => {
-
-		// Ignore root level package.json
-
-		if (packagePath === 'package.json') {
-			return true;
-		}
-
-		return !fs.existsSync(
-			path.join(path.dirname(packagePath), 'yarn.lock')
-		);
-	});
-
-	// Read package.json files
-
-	packages = await Promise.all(
-		packages.map(async (pkg) => ({
-			json: JSON.parse(await fsPromises.readFile(pkg, 'utf-8')),
-			path: pkg,
-		}))
-	);
-
-	const errors = [];
-
-	//
-	// Global checks
-	//
-
-	// Check devDependencies so that custom builds don't introduce multiple versions of a package
-
-	const allDevDependencies = packages.reduce((allDevDependencies, pkg) => {
-		const {devDependencies} = pkg.json;
-
-		if (!devDependencies) {
-			return allDevDependencies;
-		}
-
-		for (const [dependency, version] of Object.entries(devDependencies)) {
-			if (!allDevDependencies[dependency]) {
-				allDevDependencies[dependency] = {
-					paths: [],
-					versions: new Set(),
-				};
-			}
-
-			allDevDependencies[dependency].paths.push(pkg.path);
-			allDevDependencies[dependency].versions.add(version);
-		}
-
-		return allDevDependencies;
-	}, {});
-
-	const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
-
-	const nodeScriptsPackageJson = JSON.parse(
-		await fsPromises.readFile(
-			path.join(__dirname, '..', 'package.json'),
-			'utf-8'
-		)
-	);
-
-	for (const [dependency, version] of Object.entries(
-		nodeScriptsPackageJson.dependencies
-	)) {
-		if (!allDevDependencies[dependency]) {
-			allDevDependencies[dependency] = {
-				paths: [],
-				versions: new Set(),
-			};
-		}
-
-		allDevDependencies[dependency].paths.push('_node-scripts/package.json');
-		allDevDependencies[dependency].versions.add(version);
-	}
-
-	for (const [dependency, {paths, versions}] of Object.entries(
-		allDevDependencies
-	)) {
-		if (versions.size > 1) {
-			errors.push(
-				`Multiple versions for dev dependency '${dependency}' ` +
-					`have been found: ${[...versions].join(', ')}\n` +
-					`${paths.map((path) => `       - ${path}`).join('\n')}`
-			);
-		}
-	}
-
-	//
-	// Per project checks
-	//
-
-	const definedDependenciesSet = await collectDefinedDependencies();
-
-	packages.forEach((pkg) => {
-		const bad = (message) => errors.push(`${pkg.path}: BAD - ${message}`);
-
-		try {
-			const {dependencies, main, name, scripts} = pkg.json;
-
-			// Check for bad package names.
-
-			if (
-				name &&
-				!name.startsWith('@liferay/') &&
-				!ALLOWED_NAMED_SCOPE_EXCEPTIONS.includes(name)
-			) {
-				bad(
-					`package name ${name} should be under @liferay/ named scope - https://git.io/JOgy7`
-				);
-			}
-
-			// Check for banned dependencies.
-
-			const dependencyNames = dependencies
-				? Object.keys(dependencies)
-				: [];
-
-			dependencyNames.forEach((name) => {
-				if (
-					!definedDependenciesSet.has(name) &&
-					!ALLOWED_NON_GLOBAL_DEPENDENCIES.includes(name)
-				) {
-					bad(
-						`dependency not provided by a specific module: ${name} - See https://issues.liferay.com/browse/LPS-168443\n`
-					);
-				}
-			});
-
-			// Check for main entry point
-
-			const moduleDir = path.join(pkg.path, '..');
-
-			if (!main) {
-				const indexExists = [
-					'index.js',
-					'index.es.js',
-					'index.ts',
-					'index.tsx',
-				].find(
-					(file) =>
-						fs.existsSync(
-							path.join(
-								moduleDir,
-								'src/main/resources/META-INF/resources',
-								file
-							)
-						) ||
-						fs.existsSync(
-							path.join(
-								moduleDir,
-								'src/main/resources/META-INF/resources/js',
-								file
-							)
-						)
-				);
-
-				if (indexExists) {
-					bad(
-						`package.json doesn't contain a "main" entry point when you have an ${indexExists} file - https://github.com/liferay/liferay-frontend-projects/issues/719`
-					);
-				}
-			}
-
-			// Check that main entry point doesn't 'export default'
-
-			if (main && main !== 'package.json') {
-				const filePath = path.join(
-					moduleDir,
-					'src/main/resources/META-INF/resources',
-					main
-				);
-
-				if (!fs.existsSync(filePath)) {
-					bad(
-						`package.json contains a "main" entry point that doesn't exist.`
-					);
-				}
-				else {
-					const entryFile = fs.readFileSync(filePath);
-
-					if (entryFile.toString().match(/\s*export\s+default\s*/i)) {
-						bad(
-							`package.json's "main" entry point contains "export default". Use named exports only.`
-						);
-					}
-				}
-			}
-
-			// Check whether npm scripts are valid
-
-			if (scripts) {
-				for (const name of Object.keys(scripts)) {
-					if (!ALLOWED_NPM_SCRIPTS[name]) {
-						bad(
-							`package.json's script "${name}" is not allowed: ${scripts[name]}`
-						);
-					}
-					else {
-						const script = scripts[name];
-
-						let ok = false;
-
-						for (const expr of ALLOWED_NPM_SCRIPTS[name]) {
-							if (
-								(typeof expr === 'string' && script === expr) ||
-								(expr instanceof RegExp && script.match(expr))
-							) {
-								ok = true;
-								break;
-							}
-						}
-
-						if (!ok) {
-							bad(
-								`package.json's script "${name}" contains invalid value: ${scripts[name]}`
-							);
-						}
-					}
-				}
-			}
-		}
-		catch (error) {
-			bad(`error thrown during checks: ${error}`);
-		}
-	});
-
-	return errors;
-}
-
-async function collectDefinedDependencies() {
-	let rootConfig = {};
-
-	const rootDir = await getRootDir();
-
-	if (!rootDir) {
-		return new Set();
-	}
-
-	try {
-		rootConfig = await projectScopeRequire(
-			path.join(rootDir, 'node-scripts.config.js')
-		);
-	}
-	catch (error) {
-		return new Set();
-	}
-
-	const esmImports = new Set(
-		Object.entries(rootConfig.imports).reduce(
-			(acc, [moduleName, dependencies]) => {
-				return [...acc, moduleName, ...dependencies];
-			},
-			[]
-		)
-	);
-
-	return new Set([...esmImports]);
-}
-
-const ALLOWED_NAMED_SCOPE_EXCEPTIONS = [
-	'account-admin-web',
-	'adaptive-media-image-js-web',
-	'adaptive-media-web',
-	'admin-dxp-theme',
-	'analytics-client-js',
-	'analytics-reports-web',
-	'announcements-web',
-	'app-builder-web',
-	'app-builder-workflow-web',
-	'asset-categories-admin-web',
-	'asset-categories-item-selector-web',
-	'asset-list-web',
-	'asset-publisher-web',
-	'asset-taglib',
-	'asset-tags-admin-web',
-	'blogs-web',
-	'bookmarks-web',
-	'calendar-web',
-	'change-tracking-web',
-	'classic-dxp-theme',
-	'click-to-chat-web',
-	'com-liferay-dynamic-data-mapping-test',
-	'com-liferay-osb-loop-private',
-	'com.liferay.osb.www.resources',
-	'commerce-bom-admin-web',
-	'commerce-bom-web',
-	'commerce-cart-taglib',
-	'commerce-dashboard-web',
-	'commerce-frontend-impl',
-	'commerce-frontend-js',
-	'commerce-frontend-taglib',
-	'commerce-organization-web',
-	'commerce-product-content-web',
-	'commerce-product-options-web',
-	'contacts-web',
-	'content-dashboard-web',
-	'data-engine-js-components-web',
-	'data-engine-rest-impl',
-	'data-engine-taglib',
-	'depot-web',
-	'document-library-opener-onedrive-web',
-	'document-library-preview-audio',
-	'document-library-preview-document',
-	'document-library-preview-image',
-	'document-library-preview-video',
-	'document-library-video',
-	'document-library-web',
-	'dxp-cloud-emulator',
-	'dynamic-data-lists-web',
-	'dynamic-data-mapping-data-provider-web',
-	'dynamic-data-mapping-form-builder',
-	'dynamic-data-mapping-form-field-type',
-	'dynamic-data-mapping-form-renderer',
-	'dynamic-data-mapping-form-report-web',
-	'dynamic-data-mapping-form-web',
-	'dynamic-data-mapping-web',
-	'expando-web',
-	'export-import-changeset-taglib',
-	'exportimport-web',
-	'flags-taglib',
-	'forms-theme-contributor',
-	'fragment-display-web',
-	'fragment-renderer-collection-filter-impl',
-	'fragment-renderer-react-impl',
-	'fragment-resources',
-	'fragment-web',
-	'frontend-compatibility-ie',
-	'frontend-editor-alloyeditor-web',
-	'frontend-editor-ckeditor-web',
-	'frontend-image-editor-capability-brightness',
-	'frontend-image-editor-capability-contrast',
-	'frontend-image-editor-capability-crop',
-	'frontend-image-editor-capability-effects',
-	'frontend-image-editor-capability-resize',
-	'frontend-image-editor-capability-rotate',
-	'frontend-image-editor-capability-saturation',
-	'frontend-image-editor-web',
-	'frontend-js-alert-support-web',
-	'frontend-js-aui-web',
-	'frontend-js-clay-sample-web',
-	'frontend-js-collapse-support-web',
-	'frontend-js-components-web',
-	'frontend-js-dropdown-support-web',
-	'frontend-js-jquery-web',
-	'frontend-js-loader-modules-extender',
-	'frontend-js-lodash-web',
-	'frontend-js-node-shims',
-	'frontend-js-react-web',
-	'frontend-js-recharts',
-	'frontend-js-spa-web',
-	'frontend-js-svg4everybody-web',
-	'frontend-js-tabs-support-web',
-	'frontend-js-tooltip-support-web',
-	'frontend-js-web',
-	'frontend-taglib',
-	'frontend-taglib-chart',
-	'frontend-taglib-clay',
-	'frontend-taglib-clay-sample-web',
-	'frontend-taglib-clay-test-alert-toast-sample-web',
-	'frontend-theme-classic-style-guide-sample-web',
-	'frontend-theme-font-awesome-web',
-	'headless-discovery-web',
-	'hello-soy-navigation-web',
-	'hello-soy-web',
-	'hubspot-js',
-	'invitation-invite-members-web',
-	'item-selector-taglib',
-	'item-selector-upload-web',
-	'item-selector-url-web',
-	'item-selector-web',
-	'japan-theme',
-	'journal-article-dynamic-data-mapping-form-field-type',
-	'journal-web',
-	'knowledge-base-web',
-	'layout-admin-web',
-	'layout-content-page-editor-web',
-	'layout-dynamic-data-mapping-form-field-type',
-	'layout-item-selector-web',
-	'layout-reports-web',
-	'layout-seo-web',
-	'layout-set-prototype-web',
-	'layout-taglib',
-	'lfris-www-components',
-	'liferay-admin-theme',
-	'liferay-classic-theme',
-	'liferay-fjord-theme',
-	'liferay-frontend-theme-styled',
-	'liferay-frontend-theme-unstyled',
-	'liferay-learn',
-	'liferay-node-assert',
-	'liferay-node-buffer',
-	'liferay-node-console',
-	'liferay-node-constants',
-	'liferay-node-domain',
-	'liferay-node-events',
-	'liferay-node-os',
-	'liferay-node-path',
-	'liferay-node-process',
-	'liferay-node-punycode',
-	'liferay-node-querystring',
-	'liferay-node-setimmediate',
-	'liferay-node-string_decoder',
-	'liferay-node-timers',
-	'liferay-node-tty',
-	'liferay-node-url',
-	'liferay-node-util',
-	'liferay-node-vm',
-	'liferay-porygon-theme',
-	'liferay-user-dashboard-theme',
-	'liferay-user-profile-theme',
-	'liferay-watson-web',
-	'liferay-westeros-bank-theme',
-	'map-google-maps',
-	'map-openstreetmap',
-	'marketing-fragments',
-	'marketplace-store-web',
-	'message-boards-web',
-	'minium-theme',
-	'multi-factor-authentication-fido2-web',
-	'multi-factor-authentication-timebased-otp-web',
-	'my-configurable-fragment',
-	'my-sites-web',
-	'my-subscriptions-web',
-	'notifications-web',
-	'oauth2-provider-web',
-	'osb-commerce-portal-instance-admin-theme',
-	'osb-commerce-provisioning-theme',
-	'osb-commerce-provisioning-theme-impl',
-	'osb-commerce-provisioning-web',
-	'osb-community-doc-project-heading-web',
-	'osb-community-doc-project-index-web',
-	'osb-community-doc-project-random-nine-web',
-	'osb-community-github-top-contributors-web',
-	'osb-community-meetup-web',
-	'osb-community-theme',
-	'osb-customer-account-entry-details',
-	'osb-customer-downloads-display',
-	'osb-customer-release-tool',
-	'osb-customer-theme',
-	'osb-emulator',
-	'osb-events-theme',
-	'osb-faro-theme',
-	'osb-faro-web',
-	'osb-knowledge-base-theme',
-	'osb-loop-theme',
-	'osb-provisioning-theme',
-	'osb-provisioning-web',
-	'osb-www-foundations-theme-contributor',
-	'osb-www-theme',
-	'password-policies-admin-web',
-	'polls-web',
-	'portal-portlet-bridge-soy-impl',
-	'portal-reports-engine-console-web',
-	'portal-search-admin-web',
-	'portal-search-ranking-web',
-	'portal-search-synonyms-web',
-	'portal-search-web',
-	'portal-template-react-renderer-impl',
-	'portal-workflow-kaleo-designer-web',
-	'portal-workflow-kaleo-forms-web',
-	'portal-workflow-metrics-web',
-	'portal-workflow-task-web',
-	'portal-workflow-web',
-	'portlet-configuration-css-web',
-	'portlet-configuration-web',
-	'poshi-language-support',
-	'product-navigation-applications-menu-web',
-	'product-navigation-control-menu',
-	'product-navigation-control-menu-web',
-	'product-navigation-simulation-device',
-	'product-navigation-taglib',
-	'questions-web',
-	'ratings-taglib',
-	'redirect-web',
-	'remote-app-client-js',
-	'remote-app-support-web',
-	'roles-admin-web',
-	'segments-experiment-web',
-	'segments-simulation-web',
-	'segments-web',
-	'server-admin-web',
-	'sharing-taglib',
-	'sharing-web',
-	'site-admin-web',
-	'site-membership-web',
-	'site-navigation-admin-web',
-	'site-navigation-item-selector-web',
-	'site-navigation-menu-item-layout',
-	'site-navigation-menu-web',
-	'site-teams-web',
-	'social-bookmarks-taglib',
-	'speedwell-theme',
-	'staging-bar-web',
-	'staging-processes-web',
-	'staging-taglib',
-	'style-book-web',
-	'testray-theme',
-	'theme-contributor',
-	'translation-web',
-	'trash-web',
-	'user-associated-data-web',
-	'user-dashboard-dxp-theme',
-	'user-groups-admin-web',
-	'user-profile-dxp-theme',
-	'users-admin-web',
-	'watson-theme',
-	'wiki-web',
-	'youtube-web',
-];
-
-const ALLOWED_NON_GLOBAL_DEPENDENCIES = [
-	'@liferay/amd-loader',
-	'@liferay/npm-scripts',
-	'@vscode/ripgrep',
-	'alloy-ui',
-	'alloyeditor',
-	'axios',
-	'base64-js',
-	'browser-tabs-lock',
-	'ckeditor4-react',
-	'ckeditor4',
-	'codemirror',
-	'core-js',
-	'd3',
-	'es-module-shims',
-	'esbuild',
-	'fetch-mock',
-	'fs',
-	'gulp',
-	'hash.js',
-	'history',
-	'html-webpack-plugin',
-	'jest-fetch-mock',
-	'leaflet',
-	'liferay-font-awesome',
-	'liferay-theme-tasks',
-	'lodash',
-	'mini-css-extract-plugin',
-	'minimist',
-	'os-browserify',
-	'path-browserify',
-	'path-to-regexp',
-	'react-dnd-test-utils',
-	'recharts',
-	'resize-observer-polyfill',
-	'resolve',
-	'swagger-ui-react',
-	'timers-browserify',
-	'webpack',
-];
-
-const ALLOWED_NPM_SCRIPTS = {
-	'build': [
-		'liferay-npm-scripts theme build',
-		'node-scripts build',
-		'webpack',
-
-		// Format for custom builds is node scripts/buildXXX.mjs. A custom build can appear alone or
-		// after `node-scripts build`.
-
-		/^node scripts\/build.*\.mjs$/,
-		/^node-scripts build && node scripts\/build.*\.mjs$/,
-	],
-	'checkFormat': ['node-scripts format --check'],
-	'coverage': [
-
-		// Some scripts pass extra arguments.
-
-		/^liferay-npm-scripts test --coverage[^;&].*$/,
-	],
-	'format': ['node-scripts format'],
-	'start': ['webpack serve'],
-	'test': [
-
-		// Some scripts define env vars before `liferay-npm-scripts test` and others pass extra
-		// arguments.
-
-		/^[^&;]*liferay-npm-scripts test[^&;]*$/,
-	],
-	'test:watch': [
-
-		// Some scripts define env vars before `liferay-npm-scripts test --watch` and others may
-		// want to pass extra arguments.
-
-		/^[^&;]*liferay-npm-scripts test --watch[^&;]*$/,
-	],
-};
