+++ modules/_node-scripts/preflight/checkPackageNames.mjs
@@ -1,0 +1,282 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+const ALLOWED_NAMED_SCOPE_EXCEPTIONS = [
+	'account-admin-web',
+	'adaptive-media-image-js-web',
+	'adaptive-media-web',
+	'admin-dxp-theme',
+	'analytics-client-js',
+	'analytics-reports-web',
+	'announcements-web',
+	'app-builder-web',
+	'app-builder-workflow-web',
+	'asset-categories-admin-web',
+	'asset-categories-item-selector-web',
+	'asset-list-web',
+	'asset-publisher-web',
+	'asset-taglib',
+	'asset-tags-admin-web',
+	'blogs-web',
+	'bookmarks-web',
+	'calendar-web',
+	'change-tracking-web',
+	'classic-dxp-theme',
+	'click-to-chat-web',
+	'com-liferay-dynamic-data-mapping-test',
+	'com-liferay-osb-loop-private',
+	'com.liferay.osb.www.resources',
+	'commerce-bom-admin-web',
+	'commerce-bom-web',
+	'commerce-cart-taglib',
+	'commerce-dashboard-web',
+	'commerce-frontend-impl',
+	'commerce-frontend-js',
+	'commerce-frontend-taglib',
+	'commerce-organization-web',
+	'commerce-product-content-web',
+	'commerce-product-options-web',
+	'contacts-web',
+	'content-dashboard-web',
+	'data-engine-js-components-web',
+	'data-engine-rest-impl',
+	'data-engine-taglib',
+	'depot-web',
+	'document-library-opener-onedrive-web',
+	'document-library-preview-audio',
+	'document-library-preview-document',
+	'document-library-preview-image',
+	'document-library-preview-video',
+	'document-library-video',
+	'document-library-web',
+	'dxp-cloud-emulator',
+	'dynamic-data-lists-web',
+	'dynamic-data-mapping-data-provider-web',
+	'dynamic-data-mapping-form-builder',
+	'dynamic-data-mapping-form-field-type',
+	'dynamic-data-mapping-form-renderer',
+	'dynamic-data-mapping-form-report-web',
+	'dynamic-data-mapping-form-web',
+	'dynamic-data-mapping-web',
+	'expando-web',
+	'export-import-changeset-taglib',
+	'exportimport-web',
+	'flags-taglib',
+	'forms-theme-contributor',
+	'fragment-display-web',
+	'fragment-renderer-collection-filter-impl',
+	'fragment-renderer-react-impl',
+	'fragment-resources',
+	'fragment-web',
+	'frontend-compatibility-ie',
+	'frontend-editor-alloyeditor-web',
+	'frontend-editor-ckeditor-web',
+	'frontend-image-editor-capability-brightness',
+	'frontend-image-editor-capability-contrast',
+	'frontend-image-editor-capability-crop',
+	'frontend-image-editor-capability-effects',
+	'frontend-image-editor-capability-resize',
+	'frontend-image-editor-capability-rotate',
+	'frontend-image-editor-capability-saturation',
+	'frontend-image-editor-web',
+	'frontend-js-alert-support-web',
+	'frontend-js-aui-web',
+	'frontend-js-clay-sample-web',
+	'frontend-js-collapse-support-web',
+	'frontend-js-components-web',
+	'frontend-js-dropdown-support-web',
+	'frontend-js-jquery-web',
+	'frontend-js-loader-modules-extender',
+	'frontend-js-lodash-web',
+	'frontend-js-node-shims',
+	'frontend-js-react-web',
+	'frontend-js-recharts',
+	'frontend-js-spa-web',
+	'frontend-js-svg4everybody-web',
+	'frontend-js-tabs-support-web',
+	'frontend-js-tooltip-support-web',
+	'frontend-js-web',
+	'frontend-taglib',
+	'frontend-taglib-chart',
+	'frontend-taglib-clay',
+	'frontend-taglib-clay-sample-web',
+	'frontend-taglib-clay-test-alert-toast-sample-web',
+	'frontend-theme-classic-style-guide-sample-web',
+	'frontend-theme-font-awesome-web',
+	'headless-discovery-web',
+	'hello-soy-navigation-web',
+	'hello-soy-web',
+	'hubspot-js',
+	'invitation-invite-members-web',
+	'item-selector-taglib',
+	'item-selector-upload-web',
+	'item-selector-url-web',
+	'item-selector-web',
+	'japan-theme',
+	'journal-article-dynamic-data-mapping-form-field-type',
+	'journal-web',
+	'knowledge-base-web',
+	'layout-admin-web',
+	'layout-content-page-editor-web',
+	'layout-dynamic-data-mapping-form-field-type',
+	'layout-item-selector-web',
+	'layout-reports-web',
+	'layout-seo-web',
+	'layout-set-prototype-web',
+	'layout-taglib',
+	'lfris-www-components',
+	'liferay-admin-theme',
+	'liferay-classic-theme',
+	'liferay-fjord-theme',
+	'liferay-frontend-theme-styled',
+	'liferay-frontend-theme-unstyled',
+	'liferay-learn',
+	'liferay-node-assert',
+	'liferay-node-buffer',
+	'liferay-node-console',
+	'liferay-node-constants',
+	'liferay-node-domain',
+	'liferay-node-events',
+	'liferay-node-os',
+	'liferay-node-path',
+	'liferay-node-process',
+	'liferay-node-punycode',
+	'liferay-node-querystring',
+	'liferay-node-setimmediate',
+	'liferay-node-string_decoder',
+	'liferay-node-timers',
+	'liferay-node-tty',
+	'liferay-node-url',
+	'liferay-node-util',
+	'liferay-node-vm',
+	'liferay-porygon-theme',
+	'liferay-user-dashboard-theme',
+	'liferay-user-profile-theme',
+	'liferay-watson-web',
+	'liferay-westeros-bank-theme',
+	'map-google-maps',
+	'map-openstreetmap',
+	'marketing-fragments',
+	'marketplace-store-web',
+	'message-boards-web',
+	'minium-theme',
+	'multi-factor-authentication-fido2-web',
+	'multi-factor-authentication-timebased-otp-web',
+	'my-configurable-fragment',
+	'my-sites-web',
+	'my-subscriptions-web',
+	'notifications-web',
+	'oauth2-provider-web',
+	'osb-commerce-portal-instance-admin-theme',
+	'osb-commerce-provisioning-theme',
+	'osb-commerce-provisioning-theme-impl',
+	'osb-commerce-provisioning-web',
+	'osb-community-doc-project-heading-web',
+	'osb-community-doc-project-index-web',
+	'osb-community-doc-project-random-nine-web',
+	'osb-community-github-top-contributors-web',
+	'osb-community-meetup-web',
+	'osb-community-theme',
+	'osb-customer-account-entry-details',
+	'osb-customer-downloads-display',
+	'osb-customer-release-tool',
+	'osb-customer-theme',
+	'osb-emulator',
+	'osb-events-theme',
+	'osb-faro-theme',
+	'osb-faro-web',
+	'osb-knowledge-base-theme',
+	'osb-loop-theme',
+	'osb-provisioning-theme',
+	'osb-provisioning-web',
+	'osb-www-foundations-theme-contributor',
+	'osb-www-theme',
+	'password-policies-admin-web',
+	'polls-web',
+	'portal-portlet-bridge-soy-impl',
+	'portal-reports-engine-console-web',
+	'portal-search-admin-web',
+	'portal-search-ranking-web',
+	'portal-search-synonyms-web',
+	'portal-search-web',
+	'portal-template-react-renderer-impl',
+	'portal-workflow-kaleo-designer-web',
+	'portal-workflow-kaleo-forms-web',
+	'portal-workflow-metrics-web',
+	'portal-workflow-task-web',
+	'portal-workflow-web',
+	'portlet-configuration-css-web',
+	'portlet-configuration-web',
+	'poshi-language-support',
+	'product-navigation-applications-menu-web',
+	'product-navigation-control-menu',
+	'product-navigation-control-menu-web',
+	'product-navigation-simulation-device',
+	'product-navigation-taglib',
+	'questions-web',
+	'ratings-taglib',
+	'redirect-web',
+	'remote-app-client-js',
+	'remote-app-support-web',
+	'roles-admin-web',
+	'segments-experiment-web',
+	'segments-simulation-web',
+	'segments-web',
+	'server-admin-web',
+	'sharing-taglib',
+	'sharing-web',
+	'site-admin-web',
+	'site-membership-web',
+	'site-navigation-admin-web',
+	'site-navigation-item-selector-web',
+	'site-navigation-menu-item-layout',
+	'site-navigation-menu-web',
+	'site-teams-web',
+	'social-bookmarks-taglib',
+	'speedwell-theme',
+	'staging-bar-web',
+	'staging-processes-web',
+	'staging-taglib',
+	'style-book-web',
+	'testray-theme',
+	'theme-contributor',
+	'translation-web',
+	'trash-web',
+	'user-associated-data-web',
+	'user-dashboard-dxp-theme',
+	'user-groups-admin-web',
+	'user-profile-dxp-theme',
+	'users-admin-web',
+	'watson-theme',
+	'wiki-web',
+	'youtube-web',
+];
+
+export async function checkPackageNames(packageJSONFiles) {
+	const errors = [];
+
+	packageJSONFiles.forEach((pkg) => {
+		const bad = (message) => errors.push(`${pkg.path}: ${message}`);
+
+		try {
+			const {name} = pkg.json;
+
+			if (
+				name &&
+				!name.startsWith('@liferay/') &&
+				!ALLOWED_NAMED_SCOPE_EXCEPTIONS.includes(name)
+			) {
+				bad(
+					`package name ${name} should be under @liferay/ named scope - https://git.io/JOgy7`
+				);
+			}
+		}
+		catch (error) {
+			bad(`error thrown during checks: ${error}`);
+		}
+	});
+
+	return errors;
+}
